+++ src/Adapter/Authentication/AuthenticationSuccess.mjs
@@ -3,7 +3,9 @@ import { AUTHENTICATION_SUCCESS } from "./AUTHENTICATION_SUCCESS.mjs";
 //const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
 
 if (opener !== null) {
-    opener.postMessage(AUTHENTICATION_SUCCESS);
+    opener.postMessage({
+        type: AUTHENTICATION_SUCCESS
+    });
 
     close();
 } else {